Split App into View Components; Add View Routes
